Chỉnh lại price để tránh lỗi quá số lượng

+++ database/migrations/2025_12_29_090355_create_rent_posts_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
 
             $table->string('title', 200);
             $table->text('description')->nullable();
-            $table->decimal('price', 12, 2);
+            $table->bigInteger('price');
             $table->decimal('deposit', 12, 2)->nullable();
             $table->float('area')->nullable();
             $table->string('address')->nullable();
